+++ Source/ColonyManagerRedux/WorkGivers/WorkGiver_Manager.cs
@@ -39,10 +39,17 @@ internal sealed class WorkGiver_Manage : WorkGiver_Scanner
             return false;
         }
 
+        // Don't issue manager work when there's an active AI manager.
+        if (Find.CurrentMap.listerBuildings.AllBuildingsColonistOfClass<Building_AIManager>().Any(b => b.Powered))
+        {
+            JobFailReason.Is("ColonyManagerRedux.ColonyManager.CannotManage.AIManager".Translate());
+            return false;
+        }
+
         if (pawn.Dead ||
-             pawn.Downed ||
-             pawn.IsBurning() ||
-             t.IsBurning())
+            pawn.Downed ||
+            pawn.IsBurning() ||
+            t.IsBurning())
         {
             return false;
         }
@@ -53,22 +60,21 @@ internal sealed class WorkGiver_Manage : WorkGiver_Scanner
         }
 
         var power = t.TryGetComp<CompPowerTrader>();
-        if (power != null &&
-             !power.PowerOn)
+        if (power != null && !power.PowerOn)
         {
-            JobFailReason.Is("ColonyManagerRedux.CannotManage.NoPower".Translate());
+            JobFailReason.Is("ColonyManagerRedux.ColonyManager.CannotManage.NoPower".Translate());
             return false;
         }
 
         if (!Manager.For(pawn.Map).JobTracker.JobsOfType<ManagerJob>().Any())
         {
-            JobFailReason.Is("ColonyManagerRedux.CannotManage.NoJobs".Translate());
+            JobFailReason.Is("ColonyManagerRedux.ColonyManager.CannotManage.NoJobs".Translate());
             return false;
         }
 
         if (Manager.For(pawn.Map).JobTracker.NextJob == null)
         {
-            JobFailReason.Is("ColonyManagerRedux.CannotManage.NoActiveJobs".Translate());
+            JobFailReason.Is("ColonyManagerRedux.ColonyManager.CannotManage.NoActiveJobs".Translate());
             return false;
         }